refactor: スレッドアクセス時に発生しうるイベントで共通している箇所を1つのクラスにまとめる
